Fix/3.0 (#40)
* Catches IndexOutOfBounds and throws SQLException

Makes the test not dependend on pre-existing data

* Use the v3 of the engine endpoint

It uses the new updated route which relies on the SQL statement passed
as the first message on the websocket.

When retrieving a records the Jdbc driver only picks the values returned
in the `value`. Otherwise it will have to merge the Key and Value
schemas and column renaming would be required. This will make it quite
unintuitive for the user as well. A user can still grab the key value
and/or fields by doing SELECT _key as key

All tests are now creating their own data and does not rely on Lenses
having pre-existing topics.

To avoid IndexOutOfBounds found when a specific column is null the
rows.kt returns null (this needs to be revisited and improved)

Brings Kafka 2.1.0 as deps for testing

* Remove unused code

* Sets the project group to io.lenses

Renames LDriver to LensesDriver

* Sets the connection protocol to be jdbc:lenses:kafka

* Fix the adr