Wip/3.0 (#39)
* Version bumps for the 3.0 release

* Refactoring of lenses statements

* Rebuilding of result sets based on coroutine channels

* Bumped version of avro, removed joda

* Added offset row ids; some package reorg

* Implemented of new JDBC client using websockets vs coroutines!

* Updates to lenses client for websocket compatibility, introduced mappers, authentication now working over new client; implementation of fetch tables

* Completed websocket result sets, table types and tables being pulled down; parsing of lenses samsql output into rows and records; fixed schema definitions

* Converted most resultsets over to the new result set types

* Updated selects to use new client and code

* Reimplemented websocket client and wired into samsql selects

* Migrated all packages to io.lenses and removed some obsolute rest client code

* Fixed thread blocking issues in client; added tests for describe table

* Expanded out range of samsql tests in statement execute

* Loads of bugs fixed, all existing tests passing, tests added for all major statement types; all pre-samsql code removed; massive tidyup of all code

* Bumped kotlin versions

* Fix/3.0 (#40)

* Catches IndexOutOfBounds and throws SQLException

Makes the test not dependend on pre-existing data

* Use the v3 of the engine endpoint

It uses the new updated route which relies on the SQL statement passed
as the first message on the websocket.

When retrieving a records the Jdbc driver only picks the values returned
in the `value`. Otherwise it will have to merge the Key and Value
schemas and column renaming would be required. This will make it quite
unintuitive for the user as well. A user can still grab the key value
and/or fields by doing SELECT _key as key

All tests are now creating their own data and does not rely on Lenses
having pre-existing topics.

To avoid IndexOutOfBounds found when a specific column is null the
rows.kt returns null (this needs to be revisited and improved)

Brings Kafka 2.1.0 as deps for testing

* Remove unused code

* Sets the project group to io.lenses

Renames LDriver to LensesDriver

* Sets the connection protocol to be jdbc:lenses:kafka

* Fix the adr

* Updated row support to normalize values against the schema

* Added fix for deleted "null" records

* Changes the unit tests to create the topic via Lenses. This ensures
the topic details are available without introducing timing issues.